auth_center_project/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
Charm-Crypto==0.50
charset-normalizer==3.4.2
colorama==0.4.6
cors==1.0.1
crypto==1.4.1
cryptography==45.0.4
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
hypothesis==6.135.9
idna==3.10
Naked==0.1.32
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.9.0
pyparsing==2.4.0
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.4
requests-file==2.1.0
setuptools==80.9.0
shellescape==3.8.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tldextract==5.3.0
urllib3==2.4.0
zope.event==5.0
zope.interface==7.2
